Observations APIs response will have category, data_source and bioscore. contest/<contest_id>/regions/<region_id>/observations contest/<contest_id>/observations region/<region_id>/observations APIs shall return category, data_source and bioscore of observation as well.
As of now, only questagame api returns bioscore so, questagame observations will have bioscore fetched from api.
For inaturalist, if quality_grade is research then bioscore will be stored as 100 . For remaining observations, if average observations score is defined in Constant then that will be stored as bioscore else 20 will be stored as bioscore.
In Bioscore calculation for a region, we shall consider average of all region's observations bioscores.